[#16052] RESP DELEX command
Implement the DELEX command for conditional key deletion:
- DELEX key - unconditional delete (returns 1 if deleted, 0 if not found)
- DELEX key IFEQ value - delete only if value matches
- DELEX key IFNE value - delete only if value does not match
- DELEX key IFDEQ digest - delete only if hash digest matches
- DELEX key IFDNE digest - delete only if hash digest does not match

Also refactors XXH3 hash implementation to a shared utility class
for use by both DIGEST and DELEX commands.
